Refactor: moving noop to utils
