+++ cauldron/utils.py
@@ -3,9 +3,23 @@ Utils that assume what you are doing
 
 (Here be dragons)
 """
+import six
 
 
-def silence(func, x, exceptions=Exception, fail=False, ret=False):
+def cached(func):  # noqa
+    """
+    this is a noop
+    """
+    @six.wraps(func)
+    def wrapper(*args, **kwargs):
+        # TODO: add argument for message so not hardcoded for one
+        print("Cauldron could not locate the cached-property package"
+              "Please pip install the package to get the best wizardy")
+        return func(*args, **kwargs)
+    return wrapper
+
+
+def silence(func, farg, exceptions=Exception, fail=False, ret=False):
     """
     This is a wrapper that silences exceptions in a callable.
 
@@ -14,7 +28,7 @@ def silence(func, x, exceptions=Exception, fail=False, ret=False):
 
     Arguments:
         func (callable): a function in which the exception is silenced
-        x (anything): any variable that the callable requires
+        farg (anything): any variable that the callable requires
         exceptions (iterable, optional):
             list of exceptions that are to be caught.
             Default case catches Exception in general.
@@ -28,8 +42,8 @@ def silence(func, x, exceptions=Exception, fail=False, ret=False):
         if ret:
             # TODO: Refactor and move out into its own function,
             # really does not belong here.
-            return func(x)
+            return func(farg)
         else:
-            return bool(func(x))
+            return bool(func(farg))
     except exceptions:
         return not fail
